requirements.txt
ciclope[all]~=1.3.5
PyMCubes

numpy~=1.26.4
matplotlib~=3.9.2
pillow~=11.1.0
scikit-image~=0.24.0
mayavi~=4.8.2
colorama~=0.4.6
platformdirs~=3.10.0
psutil~=5.9.0
scipy~=1.15.1
tifffile~=2023.4.12
imageio~=2.33.1
pyvista~=0.45.2
meshio~=5.0.0
fpdf2~=2.8.3
optuna~=4.4.0
xgboost~=3.0.2
scikit-learn~=1.7.0
vtk~=9.4.2
tqdm~=4.67.1
numba~=0.61.2